requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
requests==2.31.0
slowapi
huggingface-hub
pydantic==2.5.0
python-dotenv